requirements.txt
pdfplumber>=0.11
pdf2image>=1.17
pillow>=10.0
pytesseract>=0.3.10
rapidfuzz>=3.0
anthropic>=0.40
openai>=1.50
google-generativeai>=0.8
